More GUI improvements and flocking interaction functionality. IT WORKS!!!
Implemented Agent generator class to manage a type of agent. Sliders for modifying behavior parameters now have functionality. Also tested if trig functions is bottleneck for performance, but turns out it is still rotation of polygon. I will try to manually change the points of poly next week, but I suspect that the existing javafx implementation is already pretty good.